+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2020-01-15
+
+- Added ParseIgnoreAttribute to allow fields to be ignored during deserialization
+- Improved tests with [ParseIgnore] attribute test
+
+**Changed**:
+- Moved CsvParser to the runtime assembly to be able to use during a project runtime
+
 ## [0.1.2] - 2020-01-08
 
 - Added missing meta files